refactor: mutualise les listes de preview des aides avec benefit-lists & benefits-list
